backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic-settings
celery[redis]
redis
openai
youtube-transcript-api
yt-dlp
httpx
python-multipart
tiktoken
pdfplumber
ebooklib
beautifulsoup4
newspaper4k
lxml
slowapi
